requirements.txt
asgiref==3.7.2
cmake==3.26.4
dill==0.3.6
Django==3.2.19
django-cors-headers==3.14.0
djangorestframework==3.14.0
dsgd @ git+https://github.com/Sergio-P/DSGD.git@7d9bb9f0b417582040a3e3a964997a1e853dd820
filelock==3.12.2
Jinja2==3.1.2
joblib==1.2.0
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
pandas==2.0.2
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.1.0
torch==2.0.1
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3